Add nested input key whitelisting
Previously it was not possible to whitelist nested keys. Whitelisting
`:foo` returned `{foo: {'bar' => 'baz'}}` when the input was `{'foo' =>
{'bar' => 'baz'}`. Now it's possible to whitelisted nested keys as well.

Example:
```
  extractor = BodyParamExtractor[:x, :y, {foo: [:bar, :baz]}]
```

This however didn't work well with coercions. Coercions are now
extracted to a separate function instead of coercing immediately in the
BodyParamExtractor / etc extractor. This also means than now we can
coerce all kind of inputs, even ConstantInput.