Log response body as a string, not an object
Previously, `RequestLogger` middleware attempted to parse response body
always, even if it wasn't valid JSON. This resulted in noisy and
misleading log messages "Failed to parse response body" in case of, for
instance, errors 403 returned by `Rack::CORS` middleware.

Since `log_response_body` setting is disabled by default anyway, logging
body as an object is probably an overkill and logging it as a plain
string would be sufficient for the most of the cases.