Fix parameter extraction when nested parameter is `nil`
If nested parameters are specified, i.e. `BodyParamExtractor[:x, {foo: [:bar]}]`,
but the input given is `{'x' => 'y', 'foo' => nil}`, then extracting the
parameters fails with the following error:

```
NoMethodError:
  undefined method `key?' for nil:NilClass
```

Which happens because `request_hash` is `nil` on the following line:

```
lib/salestation/web/extractors.rb:125
  extracted_data[filter] = request_hash[stringified_key] if request_hash.key?(stringified_key)
```

Change the extractor so that it wouldn't try to parse the nested
parameters if the root parameter is `nil`.