Make debug_message and message optional
Currently when using `.from` with Glia errors we need to provide Glia
error and legacy message even for new endpoints where legacy message is
not needed. To avoid this make `message` optional. Also make
`debug_message` optional for more errors, as it's optional for some
errors already.

Also the same applies to `errors` and `hints` from dry-validation, which
we do not need anymore if Glia error is used.

CHAN-1403