Add RSpec matchers
These make it easier to match and validate errors. Previously we defiend
similar matchers separately in projects but there were some problems:

* They were slightly different across projects
* Matching errors almost never gave any useful information
* Validating input errors was especially annoying

I also included GliaInputValidationErrorMatcher in this project. You
could argue that this should be part of `glia-errors` project, but as
this is just starting out and we need to test the matchers composition,
it is easier if they're in the same place. If the matcher gets mature
then I think we could also more it to glia-errors.