Add request input key renamer
This allows changing deprecated values before passing them to the actual
processor. If the provided key conflicts with the already existing key, it will be
possible to override that value. By default, it doesn't override as usually, you
want to use the new already updated key-value pair.

Overriding will be useful in migration scenarios when all other microservices
might not be updated to send the new attribute and the behaviour also changes
with the attribute change.