Bump `dry-validation` from 1.6.0 to 1.7.0
This resolves the warning:

```
[dry-configurable] passing a constructor as a block is deprecated and will be removed in the next major version
Provide a `constructor:` keyword argument instead
```

Likely because dry-validation 1.7 includes bumping [dry-schema to 1.8](https://github.com/dry-rb/dry-validation/blob/main/CHANGELOG.md#changed-1)
which says `(in part, to ensure dry-configurable 0.13.0 is available)`.
This warning was coming from dry-configuration so this bump seems to
fix the message.