Ensure input extractor converts input to have to symbols keys
Having a mix and match of keys being strings vs symbols causes confusion
for the developer. The Ruby preferred way is to use symbol keys instead of
strings. A couple of scenarios caused the sub-input to have
keys as strings:

* array of hashes
* content inside of a hash but not a hash

These should now have keys as symbols. This can cause breaking changes
because the keys you use to access data will have to be changed. Such
as:

```
input.fetch('foo')
=>
input.fetch(:foo)
```

The motivation for this change was to address extracting the webhook param in
engagement-api. Webhooks are specified as an array of hashes. The
previous functionality would lead to the array of hashes to have strings
for keys, while the `webhooks` param key is a symbol.

The > v1 version of dry-schema specifies that the schema must have
symbol keys. However, if the `Param`-specific value coercion is specified,
it handles keys of both string and symbol. For consistency, I think it
best to ensure keys as symbol so that we have a consistent way to fetch
data from the input.

MSG-75
MSG-71