Symbolize body parameters keys recursively
ParamExtractor class was capable of symbolizing keys only to a certain
level in an object hierarchy. After changes, the logic is capable of
doing it recursively to however many levels of hierarchy.

The change was needed because new Dry validation libraries require all
keys to be symbolic. If key is a string, the validation library does
not find it and validation fails. All keys need to be symbolized.

In particular, there is a complex nested object that needs validation
in the overseer project. The object has an array of hashes, where inside
the hash, there is another array of hashes. The deeper array was not
covered by earlier extraction logic.

As before, the “filters” used in the extractor (the list of highest
level keys of the object that is extracted) can be simply symbols, or
also a hash of symbols. E.g. the filters list is allowed to look like
[:key1, {key2: %i[sub_key1 sub_key2]]

In reality, it is not anymore needed to use hashes of symbols. The
“filters” can also look like [:key1, :key2]. All sublevels of the keys
will be symbolized anyhow, with the new recursive approach.

AUTO-541