Add module `IPAddress` and method `extract`
Module IPAddress with method extract addded as part of Audit Trail project: https://glia.atlassian.net/wiki/spaces/ENG/pages/3127214099/Audit+Trails+Audit-logger
The project requires to have available de ip from where the request to log was made.
Since obtaining the ip from the headers include a small logic, to avoid having to add this logic to every service it was decided to be added as a method in the library.

SUDO-601